fix : itemService와 LikeController부분에 itemId의 타입을 int에서 Long으로 변경

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/item/ItemService.java
@@ -31,7 +31,7 @@ public class ItemService {
     }
 	
 	// 상품 아이디로 상품 가져오기
-    public ItemEntity getItemById(int id) {
+    public ItemEntity getItemById(Long id) {
         Optional<ItemEntity> item = itemRepository.findById(id);
         return item.orElse(null);
     }
